split agent and agent group service implementation
