+++ fleet/agent_group_service.go
@@ -1,0 +1,34 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Adapted for Orb project, modifications licensed under MPL v. 2.0:
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+package fleet
+
+import (
+	"context"
+	"github.com/ns1labs/orb/pkg/errors"
+)
+
+var (
+	ErrCreateAgentGroup = errors.New("failed to create agent group")
+)
+
+func (svc fleetService) CreateAgentGroup(ctx context.Context, token string, s AgentGroup) (AgentGroup, error) {
+	mfOwnerID, err := svc.identify(token)
+	if err != nil {
+		return AgentGroup{}, err
+	}
+
+	s.MFOwnerID = mfOwnerID
+
+	err = svc.agentGroupRepository.Save(ctx, s)
+	if err != nil {
+		return AgentGroup{}, errors.Wrap(ErrCreateAgentGroup, err)
+	}
+
+	return s, nil
+}
